+++ TestBean/pom.xml
@@ -10,6 +10,7 @@
     </parent>
 
     <artifactId>TestBean</artifactId>
+    <packaging>war</packaging>
 
     <properties>
         <maven.compiler.source>19</maven.compiler.source>
