+++ pom.xml
@@ -13,7 +13,6 @@
         <module>TestBean</module>
     </modules>
 
-
     <properties>
         <maven.compiler.source>19</maven.compiler.source>
         <maven.compiler.target>19</maven.compiler.target>
