+++ src/main/java/spring/configuration/AppConfig.java
@@ -1,10 +1,11 @@
-package example.spring.configuration;
+package spring.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.core.env.Environment;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
 import org.springframework.orm.hibernate5.HibernateTransactionManager;
 import org.springframework.orm.hibernate5.LocalSessionFactoryBean;
@@ -18,6 +19,7 @@ import java.util.Properties;
 @ComponentScan(value = "example.spring")
 @EnableWebMvc
 @PropertySource("src/main/resources/app.properties")
+@EnableJpaRepositories(value = "example.spring")
 @EnableTransactionManagement
 public class AppConfig {
     private final Environment environment;
@@ -27,7 +29,7 @@ public class AppConfig {
     }
 
     @Bean
-    public LocalSessionFactoryBean sessionFactory() {
+    public LocalSessionFactoryBean entityManagerFactory() {
         LocalSessionFactoryBean sessionFactory = new LocalSessionFactoryBean();
         sessionFactory.setDataSource(dataSource());
         sessionFactory.setPackagesToScan("example.spring.model");
@@ -38,11 +40,15 @@ public class AppConfig {
     @Bean
     public DataSource dataSource() {
         DriverManagerDataSource dataSource = new DriverManagerDataSource();
-        dataSource.setDriverClassName(environment.getRequiredProperty("db.driverClassName"));
-        dataSource.setUrl(environment.getRequiredProperty("db.url"));
-        dataSource.setUsername(environment.getRequiredProperty("db.username"));
-        dataSource.setPassword(environment.getRequiredProperty("db.password"));
-        return dataSource;
+        try {
+            dataSource.setDriverClassName(environment.getRequiredProperty("db.driverClassName"));
+            dataSource.setUrl(environment.getRequiredProperty("db.url"));
+            dataSource.setUsername(environment.getRequiredProperty("db.username"));
+            dataSource.setPassword(environment.getRequiredProperty("db.password"));
+            return dataSource;
+        } catch (Exception e) {
+            throw new RuntimeException("Failed to configure DataSource", e);
+        }
     }
 
     private Properties hibernateProperties() {
@@ -54,9 +60,9 @@ public class AppConfig {
         return properties;
     }
     @Bean
-    public HibernateTransactionManager getTransactionManager() {
+    public HibernateTransactionManager transactionManager() {
         HibernateTransactionManager transactionManager = new HibernateTransactionManager();
-        transactionManager.setSessionFactory(sessionFactory().getObject());
+        transactionManager.setSessionFactory(entityManagerFactory().getObject());
         return transactionManager;
     }
 }
